web/te-backend/requirements.txt
gunicorn~=19.7
pydub~=0.23
gevent~=1.4
django~=1.11.29
djangorestframework~=3.10.3
django-cors-headers~=3.2
coreapi~=2.3
urllib3~=1.25
whitenoise~=5.0
psycopg2-binary~=2.8
raven~=6.10
celery~=4.4
redis~=3.4
flower~=0.9
drf-yasg~=1.12
rest-social-auth~=3.0
pyyaml
